@@ -3,6 +3,7 @@ import {ChatInputCommandInteraction, EmbedBuilder, PermissionFlagsBits} from 'di
 import {injectable} from 'inversify';
 import {prisma} from '../utils/db.js';
 import Command from './index.js';
+import {getGuildSettings} from '../utils/get-guild-settings.js';
 
 @injectable()
 export default class implements Command {
@@ -33,10 +34,36 @@ export default class implements Command {
         .setDescription('whether to leave when everyone else leaves')
         .setRequired(true)))
     .addSubcommand(subcommand => subcommand
+      .setName('set-queue-add-response-hidden')
+      .setDescription('set whether bot responses to queue additions are only displayed to the requester')
+      .addBooleanOption(option => option
+        .setName('value')
+        .setDescription('whether bot responses to queue additions are only displayed to the requester')
+        .setRequired(true)))
+    .addSubcommand(subcommand => subcommand
+      .setName('set-auto-announce-next-song')
+      .setDescription('set whether to announce the next song in the queue automatically')
+      .addBooleanOption(option => option
+        .setName('value')
+        .setDescription('whether to announce the next song in the queue automatically')
+        .setRequired(true)))
+    .addSubcommand(subcommand => subcommand
+      .setName('set-default-volume')
+      .setDescription('set default volume used when entering the voice channel')
+      .addIntegerOption(option => option
+        .setName('level')
+        .setDescription('volume percentage (0 is muted, 100 is max & default)')
+        .setMinValue(0)
+        .setMaxValue(100)
+        .setRequired(true)))
+    .addSubcommand(subcommand => subcommand
       .setName('get')
       .setDescription('show all settings'));
 
   async execute(interaction: ChatInputCommandInteraction) {
+    // Ensure guild settings exist before trying to update
+    await getGuildSettings(interaction.guild!.id);
+
     switch (interaction.options.getSubcommand()) {
       case 'set-playlist-limit': {
         const limit: number = interaction.options.getInteger('limit')!;
@@ -93,14 +120,61 @@ export default class implements Command {
         break;
       }
 
+      case 'set-queue-add-response-hidden': {
+        const value = interaction.options.getBoolean('value')!;
+
+        await prisma.setting.update({
+          where: {
+            guildId: interaction.guild!.id,
+          },
+          data: {
+            queueAddResponseEphemeral: value,
+          },
+        });
+
+        await interaction.reply('👍 queue add notification setting updated');
+
+        break;
+      }
+
+      case 'set-auto-announce-next-song': {
+        const value = interaction.options.getBoolean('value')!;
+
+        await prisma.setting.update({
+          where: {
+            guildId: interaction.guild!.id,
+          },
+          data: {
+            autoAnnounceNextSong: value,
+          },
+        });
+
+        await interaction.reply('👍 auto announce setting updated');
+
+        break;
+      }
+
+      case 'set-default-volume': {
+        const value = interaction.options.getInteger('level')!;
+
+        await prisma.setting.update({
+          where: {
+            guildId: interaction.guild!.id,
+          },
+          data: {
+            defaultVolume: value,
+          },
+        });
+
+        await interaction.reply('👍 volume setting updated');
+
+        break;
+      }
+
       case 'get': {
         const embed = new EmbedBuilder().setTitle('Config');
 
-        const config = await prisma.setting.findUnique({where: {guildId: interaction.guild!.id}});
-
-        if (!config) {
-          throw new Error('no config found');
-        }
+        const config = await getGuildSettings(interaction.guild!.id);
 
         const settingsToShow = {
           'Playlist Limit': config.playlistLimit,
@@ -108,6 +182,9 @@ export default class implements Command {
             ? 'never leave'
             : `${config.secondsToWaitAfterQueueEmpties}s`,
           'Leave if there are no listeners': config.leaveIfNoListeners ? 'yes' : 'no',
+          'Auto announce next song in queue': config.autoAnnounceNextSong ? 'yes' : 'no',
+          'Add to queue reponses show for requester only': config.autoAnnounceNextSong ? 'yes' : 'no',
+          'Default Volume': config.defaultVolume,
         };
 
         let description = '';
