@@ -3,7 +3,7 @@ import {SlashCommandBuilder} from '@discordjs/builders';
 import {TYPES} from '../types.js';
 import {inject, injectable} from 'inversify';
 import PlayerManager from '../managers/player.js';
-import Command from '.';
+import Command from './index.js';
 
 @injectable()
 export default class implements Command {
@@ -28,6 +28,6 @@ export default class implements Command {
 
     player.disconnect();
 
-    await interaction.reply('u betcha');
+    await interaction.reply('u betcha, disconnected');
   }
 }
