@@ -3,7 +3,7 @@ import {SlashCommandBuilder} from '@discordjs/builders';
 import {inject, injectable} from 'inversify';
 import {TYPES} from '../types.js';
 import PlayerManager from '../managers/player.js';
-import Command from '.';
+import Command from './index.js';
 import {buildQueueEmbed} from '../utils/build-embed.js';
 
 @injectable()
