@@ -9,10 +9,12 @@ import Prisma from '@prisma/client';
 import ora from 'ora';
 import {startBot} from '../index.js';
 import logBanner from '../utils/log-banner.js';
-import {createDatabasePath} from '../utils/create-database-url.js';
+import createDatabaseUrl, {createDatabasePath} from '../utils/create-database-url.js';
 import {DATA_DIR} from '../services/config.js';
 
 const client = new Prisma.PrismaClient();
+
+process.env.DATABASE_URL = process.env.DATABASE_URL ?? createDatabaseUrl(DATA_DIR);
 
 const migrateFromSequelizeToPrisma = async () => {
   await execa('prisma', ['migrate', 'resolve', '--applied', '20220101155430_migrate_from_sequelize'], {preferLocal: true});
