@@ -1,14 +1,14 @@
 import {VoiceChannel, Snowflake} from 'discord.js';
 import {Readable} from 'stream';
 import hasha from 'hasha';
-import ytdl from 'ytdl-core';
+import ytdl, {videoFormat} from '@distube/ytdl-core';
 import {WriteStream} from 'fs-capacitor';
 import ffmpeg from 'fluent-ffmpeg';
 import shuffle from 'array-shuffle';
 import {
   AudioPlayer,
   AudioPlayerState,
-  AudioPlayerStatus,
+  AudioPlayerStatus, AudioResource,
   createAudioPlayer,
   createAudioResource, DiscordGatewayAdapterCreator,
   joinVoiceChannel,
@@ -18,7 +18,8 @@ import {
 } from '@discordjs/voice';
 import FileCacheProvider from './file-cache.js';
 import debug from '../utils/debug.js';
-import {prisma} from '../utils/db.js';
+import {getGuildSettings} from '../utils/get-guild-settings.js';
+import {buildPlayingMessageEmbed} from '../utils/build-embed.js';
 
 export enum MediaSource {
   Youtube,
@@ -33,7 +34,7 @@ export interface QueuedPlaylist {
 export interface SongMetadata {
   title: string;
   artist: string;
-  url: string;
+  url: string; // For YT, it's the video ID (not the full URI)
   length: number;
   offset: number;
   playlist: QueuedPlaylist | null;
@@ -56,15 +57,23 @@ export interface PlayerEvents {
   statusChange: (oldStatus: STATUS, newStatus: STATUS) => void;
 }
 
+type YTDLVideoFormat = videoFormat & {loudnessDb?: number};
+
+export const DEFAULT_VOLUME = 100;
+
 export default class {
   public voiceConnection: VoiceConnection | null = null;
   public status = STATUS.PAUSED;
   public guildId: string;
   public loopCurrentSong = false;
-
+  public loopCurrentQueue = false;
+  private currentChannel: VoiceChannel | undefined;
   private queue: QueuedSong[] = [];
   private queuePosition = 0;
   private audioPlayer: AudioPlayer | null = null;
+  private audioResource: AudioResource | null = null;
+  private volume?: number;
+  private defaultVolume: number = DEFAULT_VOLUME;
   private nowPlaying: QueuedSong | null = null;
   private playPositionInterval: NodeJS.Timeout | undefined;
   private lastSongURL = '';
@@ -79,10 +88,33 @@ export default class {
   }
 
   async connect(channel: VoiceChannel): Promise<void> {
+    // Always get freshest default volume setting value
+    const settings = await getGuildSettings(this.guildId);
+    const {defaultVolume = DEFAULT_VOLUME} = settings;
+    this.defaultVolume = defaultVolume;
+
     this.voiceConnection = joinVoiceChannel({
       channelId: channel.id,
       guildId: channel.guild.id,
       adapterCreator: channel.guild.voiceAdapterCreator as DiscordGatewayAdapterCreator,
+    });
+
+    // Workaround to disable keepAlive
+    this.voiceConnection.on('stateChange', (oldState, newState) => {
+      /* eslint-disable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call */
+      const oldNetworking = Reflect.get(oldState, 'networking');
+      const newNetworking = Reflect.get(newState, 'networking');
+
+      const networkStateChangeHandler = (_: any, newNetworkState: any) => {
+        const newUdp = Reflect.get(newNetworkState, 'udp');
+        clearInterval(newUdp?.keepAliveInterval);
+      };
+
+      oldNetworking?.off('stateChange', networkStateChangeHandler);
+      newNetworking?.on('stateChange', networkStateChangeHandler);
+      /* eslint-enable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call */
+
+      this.currentChannel = channel;
     });
   }
 
@@ -94,10 +126,11 @@ export default class {
 
       this.loopCurrentSong = false;
       this.voiceConnection.destroy();
-      this.audioPlayer?.stop();
+      this.audioPlayer?.stop(true);
 
       this.voiceConnection = null;
       this.audioPlayer = null;
+      this.audioResource = null;
     }
   }
 
@@ -133,9 +166,7 @@ export default class {
       },
     });
     this.voiceConnection.subscribe(this.audioPlayer);
-    this.audioPlayer.play(createAudioResource(stream, {
-      inputType: StreamType.WebmOpus,
-    }));
+    this.playAudioPlayerResource(this.createAudioStream(stream));
     this.attachListeners();
     this.startTrackingPosition(positionSeconds);
 
@@ -198,11 +229,7 @@ export default class {
         },
       });
       this.voiceConnection.subscribe(this.audioPlayer);
-      const resource = createAudioResource(stream, {
-        inputType: StreamType.WebmOpus,
-      });
-
-      this.audioPlayer.play(resource);
+      this.playAudioPlayerResource(this.createAudioStream(stream));
 
       this.attachListeners();
 
@@ -253,14 +280,10 @@ export default class {
       if (this.getCurrent() && this.status !== STATUS.PAUSED) {
         await this.play();
       } else {
-        this.audioPlayer?.stop();
         this.status = STATUS.IDLE;
+        this.audioPlayer?.stop(true);
 
-        const settings = await prisma.setting.findUnique({where: {guildId: this.guildId}});
-
-        if (!settings) {
-          throw new Error('Could not find settings for guild');
-        }
+        const settings = await getGuildSettings(this.guildId);
 
         const {secondsToWaitAfterQueueEmpties} = settings;
         if (secondsToWaitAfterQueueEmpties !== 0) {
@@ -390,36 +413,45 @@ export default class {
     return this.queue[this.queuePosition + to];
   }
 
+  setVolume(level: number): void {
+    // Level should be a number between 0 and 100 = 0% => 100%
+    this.volume = level;
+    this.setAudioPlayerVolume(level);
+  }
+
+  getVolume(): number {
+    // Only use default volume if player volume is not already set (in the event of a reconnect we shouldn't reset)
+    return this.volume ?? this.defaultVolume;
+  }
+
   private getHashForCache(url: string): string {
     return hasha(url);
   }
 
   private async getStream(song: QueuedSong, options: {seek?: number; to?: number} = {}): Promise<Readable> {
-    if (song.source === MediaSource.HLS) {
-      return this.createReadStream(song.url);
+    if (this.status === STATUS.PLAYING) {
+      this.audioPlayer?.stop();
+    } else if (this.status === STATUS.PAUSED) {
+      this.audioPlayer?.stop(true);
     }
 
-    let ffmpegInput = '';
+    if (song.source === MediaSource.HLS) {
+      return this.createReadStream({url: song.url, cacheKey: song.url});
+    }
+
+    let ffmpegInput: string | null;
     const ffmpegInputOptions: string[] = [];
     let shouldCacheVideo = false;
 
-    let format: ytdl.videoFormat | undefined;
+    let format: YTDLVideoFormat | undefined;
 
-    try {
-      ffmpegInput = await this.fileCache.getPathFor(this.getHashForCache(song.url));
+    ffmpegInput = await this.fileCache.getPathFor(this.getHashForCache(song.url));
 
-      if (options.seek) {
-        ffmpegInputOptions.push('-ss', options.seek.toString());
-      }
-
-      if (options.to) {
-        ffmpegInputOptions.push('-to', options.to.toString());
-      }
-    } catch {
+    if (!ffmpegInput) {
       // Not yet cached, must download
       const info = await ytdl.getInfo(song.url);
 
-      const {formats} = info;
+      const formats = info.formats as YTDLVideoFormat[];
 
       const filter = (format: ytdl.videoFormat): boolean => format.codecs === 'opus' && format.container === 'webm' && format.audioSampleRate !== undefined && parseInt(format.audioSampleRate, 10) === 48000;
 
@@ -453,11 +485,15 @@ export default class {
         }
       }
 
+      debug('Using format', format);
+
       ffmpegInput = format.url;
 
       // Don't cache livestreams or long videos
       const MAX_CACHE_LENGTH_SECONDS = 30 * 60; // 30 minutes
       shouldCacheVideo = !info.player_response.videoDetails.isLiveContent && parseInt(info.videoDetails.lengthSeconds, 10) < MAX_CACHE_LENGTH_SECONDS && !options.seek;
+
+      debug(shouldCacheVideo ? 'Caching video' : 'Not caching video');
 
       ffmpegInputOptions.push(...[
         '-reconnect',
@@ -467,17 +503,23 @@ export default class {
         '-reconnect_delay_max',
         '5',
       ]);
-
-      if (options.seek) {
-        ffmpegInputOptions.push('-ss', options.seek.toString());
-      }
-
-      if (options.to) {
-        ffmpegInputOptions.push('-to', options.to.toString());
-      }
     }
 
-    return this.createReadStream(ffmpegInput, {ffmpegInputOptions, cache: shouldCacheVideo});
+    if (options.seek) {
+      ffmpegInputOptions.push('-ss', options.seek.toString());
+    }
+
+    if (options.to) {
+      ffmpegInputOptions.push('-to', options.to.toString());
+    }
+
+    return this.createReadStream({
+      url: ffmpegInput,
+      cacheKey: song.url,
+      ffmpegInputOptions,
+      cache: shouldCacheVideo,
+      volumeAdjustment: format?.loudnessDb ? `${-format.loudnessDb}dB` : undefined,
+    });
   }
 
   private startTrackingPosition(initalPosition?: number): void {
@@ -529,28 +571,48 @@ export default class {
       return;
     }
 
+    // Automatically re-add current song to queue
+    if (this.loopCurrentQueue && newState.status === AudioPlayerStatus.Idle && this.status === STATUS.PLAYING) {
+      const currentSong = this.getCurrent();
+
+      if (currentSong) {
+        this.add(currentSong);
+      } else {
+        throw new Error('No song currently playing.');
+      }
+    }
+
     if (newState.status === AudioPlayerStatus.Idle && this.status === STATUS.PLAYING) {
       await this.forward(1);
+      // Auto announce the next song if configured to
+      const settings = await getGuildSettings(this.guildId);
+      const {autoAnnounceNextSong} = settings;
+      if (autoAnnounceNextSong && this.currentChannel) {
+        await this.currentChannel.send({
+          embeds: this.getCurrent() ? [buildPlayingMessageEmbed(this)] : [],
+        });
+      }
     }
   }
 
-  private async createReadStream(url: string, options: {ffmpegInputOptions?: string[]; cache?: boolean} = {}): Promise<Readable> {
+  private async createReadStream(options: {url: string; cacheKey: string; ffmpegInputOptions?: string[]; cache?: boolean; volumeAdjustment?: string}): Promise<Readable> {
     return new Promise((resolve, reject) => {
       const capacitor = new WriteStream();
 
       if (options?.cache) {
-        const cacheStream = this.fileCache.createWriteStream(this.getHashForCache(url));
+        const cacheStream = this.fileCache.createWriteStream(this.getHashForCache(options.cacheKey));
         capacitor.createReadStream().pipe(cacheStream);
       }
 
       const returnedStream = capacitor.createReadStream();
       let hasReturnedStreamClosed = false;
 
-      const stream = ffmpeg(url)
+      const stream = ffmpeg(options.url)
         .inputOptions(options?.ffmpegInputOptions ?? ['-re'])
         .noVideo()
         .audioCodec('libopus')
         .outputFormat('webm')
+        .addOutputOption(['-filter:a', `volume=${options?.volumeAdjustment ?? '1'}`])
         .on('error', error => {
           if (!hasReturnedStreamClosed) {
             reject(error);
@@ -563,11 +625,34 @@ export default class {
       stream.pipe(capacitor);
 
       returnedStream.on('close', () => {
-        stream.kill('SIGKILL');
+        if (!options.cache) {
+          stream.kill('SIGKILL');
+        }
+
         hasReturnedStreamClosed = true;
       });
 
       resolve(returnedStream);
     });
+  }
+
+  private createAudioStream(stream: Readable) {
+    return createAudioResource(stream, {
+      inputType: StreamType.WebmOpus,
+      inlineVolume: true,
+    });
+  }
+
+  private playAudioPlayerResource(resource: AudioResource) {
+    if (this.audioPlayer !== null) {
+      this.audioResource = resource;
+      this.setAudioPlayerVolume();
+      this.audioPlayer.play(this.audioResource);
+    }
+  }
+
+  private setAudioPlayerVolume(level?: number) {
+    // Audio resource expects a float between 0 and 1 to represent level percentage
+    this.audioResource?.volume?.setVolume((level ?? this.getVolume()) / 100);
   }
 }
